+++ basic/src/main/java/com/ke/kob/basic/model/JobType.java
