+++ basic/src/main/java/com/ke/kob/basic/model/LogLevel.java
