+++ basic/src/main/java/com/ke/kob/basic/model/TaskResult.java
