+++ docs/comparison_of_products.md
